Refactor process_data and remove duplication
Moved the process_data method to process_query_results.py for better modularization and removed the duplicated method from biorxiv_api.py. This refactoring enhances code maintainability and readability.
